refactor: move exception handlers to rest controller cs-1
